blockchain: Move UTXO DB tests to separate file.
This moves the tests that deal with the UTXO database buckets from
chainio_test.go to a separate file, utxoio_test.go.

This is part of moving the UTXO set and state to a separate database.
